feat: 🎸 added support for stockfish chess engine
